refactor: consolidate 6 internal hooks into useChartCore + event-utils
Rearchitect the hook decomposition based on actual coupling boundaries:

- Merge useInstanceLifecycle + useOptionSync + useEvents into useChartCore
  which owns all shared state (refs, lastAppliedRef, boundEventsRef) internally
- Replace 16-parameter function with 3-parameter signature (ref, shouldInit, config)
- Extract bind/unbindEvents into pure event-utils.ts (no hook, no mixed exports)
- Inline trivial loading/group effects back into orchestrator (~7 lines each)
- Delete log-error.ts and types.ts (absorbed into useChartCore)
- Result: 3 internal files instead of 8, ref-sync colocated with its consumers
